refactor(db): slim StorageEngine to withConn/closeAll (move vault deletion into data layer as one tx), drop storage.ts shim and dead metrics.timed
