--- backend/src/db/storage.ts
@@ -1,7 +1,0 @@
-export type {
-	DbEngineKind,
-	EngineConfig,
-	Storage,
-	StorageEngine,
-} from "./engine.ts";
-export { createEngine } from "./engine.ts";
